fix: show available products on page
There was an error in the code that prevented the products from being displayed on the page in a production environment. This was due to Next.js caching the page and not updating it when the products were fetched from the server. This commit fixes that issue by disabling the caching of the page.